refactor(user): move `UserService.login` to `AuthService.createToken` (#207)
The UserService should only concern itself with user-related admin
tasks. Hence, move `login()`, which creates signed JWT tokens, to
AuthService, where it should be. This allows us to simplify our
dependencies by having everything JWT-related in AuthService, the
authoritative service for handling authentication
