+++ server/src/modules/post/__tests__/post.controller.spec.ts
@@ -7,6 +7,7 @@ import { PostController } from '../post.controller'
 describe('PostController', () => {
   const path = '/posts'
   const authService = {
+    createToken: jest.fn(),
     getUserIdFromToken: jest.fn(),
     getOfficerUser: jest.fn(),
     checkIfWhitelistedOfficer: jest.fn(),
